pyproject.toml
[project]
name = "desafiobackend-lacrei-saude"
version = "0.1.0"
description = "Desafio Técnico BackEnd"
authors = [
    {name = "thiago",email = "thiagoc.barros92@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "django-role-permissions (>=3.2.0,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
